기타: 패키지 변경
* 도메인 변경에 따른 패키지 변경:
  io.myrecipes.front -> link.myrecipes.front
