+++ pom.xml
@@ -8,7 +8,7 @@
         <version>2.1.7.RELEASE</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>io.myrecipes.front</groupId>
+    <groupId>link.myrecipes.front</groupId>
     <artifactId>myrecipes-front</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>myrecipes-front</name>
