--- src/main/java/io/myrecipes/front/common/S3Helper.java
@@ -1,9 +1,0 @@
-package io.myrecipes.front.common;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-public interface S3Helper {
-    String upload(MultipartFile multipartFile, String path) throws IOException;
-}
