--- src/main/java/io/myrecipes/front/config/AWSConfiguration.java
@@ -1,34 +1,0 @@
-package io.myrecipes.front.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
-
-@Configuration
-public class AWSConfiguration {
-    @Value("${cloud.aws.credentials.accessKey}")
-    private String accessKey;
-
-    @Value("${cloud.aws.credentials.secretKey}")
-    private String secretKey;
-
-    @Value("${cloud.aws.region.static}")
-    private String region;
-
-    @Bean
-    public AwsBasicCredentials awsBasicCredentials() {
-        return AwsBasicCredentials.create(accessKey, secretKey);
-    }
-
-    @Bean
-    public S3Client sClient(AwsBasicCredentials awsBasicCredentials) {
-        return S3Client.builder()
-                .credentialsProvider(StaticCredentialsProvider.create(awsBasicCredentials))
-                .region(Region.of(region))
-                .build();
-    }
-}
