--- src/main/java/io/myrecipes/front/controller/SystemController.java
@@ -1,14 +1,0 @@
-package io.myrecipes.front.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class SystemController {
-    @GetMapping("/health")
-    public String health(Model model) {
-        model.addAttribute("title", "Hello System");
-        return "health";
-    }
-}
