--- src/main/java/io/myrecipes/front/dto/RecipeTag.java
@@ -1,19 +1,0 @@
-package io.myrecipes.front.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@NoArgsConstructor
-public class RecipeTag {
-    private String tag;
-
-    @Builder
-    public RecipeTag(String tag) {
-        this.tag = tag;
-    }
-}
