+++ src/main/java/link/myrecipes/front/common/RestTemplateHelper.java
@@ -1,0 +1,11 @@
+package link.myrecipes.front.common;
+
+import java.util.List;
+
+public interface RestTemplateHelper {
+    <T> T getForEntity(Class<T> clazz, String url, Object... uriVariables);
+
+    <T> List<T> getForList(Class<T> clazz, String url, Object... uriVariables);
+
+    <T, R> T postForEntity(Class<T> clazz, String url, R body, Object... uriVariables);
+}
