+++ src/main/java/link/myrecipes/front/common/S3Helper.java
@@ -1,0 +1,9 @@
+package link.myrecipes.front.common;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+public interface S3Helper {
+    String upload(MultipartFile multipartFile, String path) throws IOException;
+}
